@@ -1,6 +1,6 @@
 {
   "name": "livekit-react",
-  "version": "0.7.1",
+  "version": "0.8.0",
   "description": "LiveKit SDK & Components",
   "author": "davidzhao",
   "license": "Apache-2.0",
@@ -25,21 +25,21 @@
     "deploy": "gh-pages -d example/build"
   },
   "peerDependencies": {
-    "react": "^15.0.0 || ^16.0.0 || ^17.0.0",
-    "react-dom": "^15.0.0 || ^16.0.0 || ^17.0.0",
-    "livekit-client": "^0.17.1"
+    "livekit-client": "^0.18.3",
+    "react": ">=15",
+    "react-dom": ">=15"
   },
   "devDependencies": {
+    "@fortawesome/fontawesome-common-types": "^0.2.36",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "@types/jest": "^25.1.4",
     "@types/node": "^12.12.38",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.13",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "@typescript-eslint/eslint-plugin": "^3.10.1",
     "@typescript-eslint/parser": "^3.10.1",
-    "@fortawesome/fontawesome-common-types": "^0.2.36",
     "babel-eslint": "^10.0.3",
     "cross-env": "^7.0.2",
     "eslint": "^7.11.0",
@@ -52,14 +52,14 @@
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-standard": "^4.0.1",
     "gh-pages": "^2.2.0",
+    "livekit-client": "^0.18.3",
     "microbundle-crl": "^0.13.10",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "livekit-client": "^0.17.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-scripts": "^4.0.3",
-    "typescript": "~4.2.3"
+    "typescript": "^4.2.3"
   },
   "files": [
     "dist",
@@ -68,10 +68,13 @@
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "^1.2.35",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.16",
+    "@fortawesome/react-fontawesome": "^0.1.18",
     "@types/react-responsive": "^8.0.2",
     "react-aspect-ratio": "^1.0.50",
     "react-responsive": "^8.2.0",
     "react-tiny-popover": "^6.0.5"
+  },
+  "resolutions": {
+    "@fortawesome/fontawesome-common-types": "0.2.36"
   }
 }
